Update single_gene_score.nf
updated params so there are "General", "TCGA" and "GTEx" sets (maybe in the future we want these to be organized in groups like params.predixcan, params.magma, ...etc)

europfile was changed to the tcga one as this is solely used for LD determination and this is a large-ish ~8k cohort of single ancestry. Previously this was set to flip between TCGA and GTEx cohorts, but GTEx is too small. We can try to use other non-cancer cohorts if those become available